requirements.txt
streamlit
openai
psycopg2-binary
sentence-transformers
pandas
numpy
